requirements.txt
asgiref==3.2.10
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.3
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1.1
django-admin-rangefilter==0.6.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-modeltranslation==0.15.2
djangorestframework==3.12.1
html5lib==1.1
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2020.1
reportlab==3.5.51
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
webencodings==0.5.1
xhtml2pdf==0.2.4
